+++ core/data/src/main/java/com/kusitms/connectdog/core/data/repository/HomeRepository.kt
@@ -6,5 +6,7 @@ import com.kusitms.connectdog.core.model.Review
 interface HomeRepository {
     suspend fun getAnnouncementList(): List<Announcement>
 
+    suspend fun getAnnouncementListWithFilter(): List<Announcement>
+
     suspend fun getReviewList(): List<Review>
 }
