+++ core/data/src/main/java/com/kusitms/connectdog/core/data/repository/HomeRepositoryImpl.kt
@@ -12,6 +12,10 @@ internal class HomeRepositoryImpl @Inject constructor(
         TODO("Not yet implemented")
     }
 
+    override suspend fun getAnnouncementListWithFilter(): List<Announcement> {
+        TODO("Not yet implemented")
+    }
+
     override suspend fun getReviewList(): List<Review> {
         TODO("Not yet implemented")
     }
